% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon.R
\docType{data}
\name{arnold1949}
\alias{arnold1949}
\title{Arnold and Libby's Curve of Knowns}
\format{
A \code{data.frame} with 6 observations and 8 variables:
\describe{
\item{sample}{\code{\link{character}}: sample name.}
\item{age_expected}{\code{\link{integer}}: Expected age (year BP).}
\item{age_expected_error}{\code{\link{integer}}: Error on \code{age_expected} (year BP).}
\item{age_found}{\code{\link{integer}}: Measured age (year BP).}
\item{age_found_error}{\code{\link{integer}}: Error on \code{age_found} (year BP).}
\item{activity_expected}{\code{\link{numeric}}: Expected specific activity (cpm/g of carbon).}
\item{activity_found}{\code{\link{numeric}}: Measured specific activity (cpm/g of carbon).}
\item{activity_found_error}{\code{\link{numeric}}: Error on \code{activity_found} (cpm/g of carbon).}
}
}
\source{
Arnold, J. R. and Libby, W. F. (1949). Age Determinations by Radiocarbon
Content: Checks with Samples of Known Age. \emph{Science}, 110(2869),
678-80. \doi{10.1126/science.110.2869.678}
}
\usage{
arnold1949
}
\description{
"The agreement between prediction and observation is seen to be
satisfactory."
}
\seealso{
Other radiocarbon dating: 
\code{\link{intcal09}},
\code{\link{intcal13}},
\code{\link{intcal20}}
}
\concept{radiocarbon dating}
\keyword{datasets}
