% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\docType{data}
\name{birds}
\alias{birds}
\title{European Birds}
\format{
A data frame with 35 rows (species) and 3 variables (woodlands).
}
\source{
Magurran, A. E. (1988). \emph{Ecological Diversity and its Measurement}.
Princeton, NJ: Princeton University Press.
\doi{10.1007/978-94-015-7358-0}.
}
\usage{
birds
}
\description{
A dataset of birds species abundance in remote European woodlands.
}
\seealso{
Other count data: 
\code{\link{boves}},
\code{\link{chevelon}},
\code{\link{compiegne}},
\code{\link{merzbach}},
\code{\link{mississippi}},
\code{\link{zuni}}
}
\concept{count data}
\keyword{datasets}
