% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\docType{data}
\name{boves}
\alias{boves}
\title{Boves Ceramics}
\format{
A \code{data.frame} with 8 observations (periods) and 13 variables (ceramic types).
}
\source{
Racinet P. (2002). Le site castral et prioral de Boves du Xe au XVIIe
siècle. Bilan des recherches 1996-2000. \emph{Revue archéologique de Picardie}.
Numéro spécial 20, 123 p.
}
\usage{
boves
}
\description{
A dataset containing the ceramic counts from the castle site of Boves
(Somme, France). The data are grouped into eight periods ranging from the
10th to the 18th century and thirteen ceramic types.
}
\seealso{
Other count data: 
\code{\link{birds}},
\code{\link{chevelon}},
\code{\link{compiegne}},
\code{\link{loire}},
\code{\link{merzbach}},
\code{\link{mississippi}},
\code{\link{zuni}}
}
\concept{count data}
\keyword{datasets}
