% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\docType{data}
\name{chevelon}
\alias{chevelon}
\title{Chevelon Ground Stone}
\format{
A \code{data.frame} with 12 observations (sites) and 10 variables (ground stone types):
\describe{
\item{BMe}{\code{\link{integer}}: basin metate.}
\item{SMe}{\code{\link{integer}}: slab metate.}
\item{TMe}{\code{\link{integer}}: Trough metate.}
\item{IMe}{\code{\link{integer}}: indeterminate metate.}
\item{UMa}{\code{\link{integer}}: unifacial mano.}
\item{BMa}{\code{\link{integer}}: bifacial mano.}
\item{MUHa}{\code{\link{integer}}: modified unifacial handstone.}
\item{MBHa}{\code{\link{integer}}: modified bifacial handstone.}
\item{UUHa}{\code{\link{integer}}: unmodified unifacial handstone.}
\item{UBHa}{\code{\link{integer}}: unmodified bifacial handstone.}
}
}
\source{
Reid, J. J. (ed.) (1982). \emph{Cholla Project Archaeology}. Vol. 2.
Archaeological Series 161. Tucson: University of Arizona.
\doi{10.6067/XCV8435710}
}
\usage{
chevelon
}
\description{
A dataset of ground stone artifact counts from the Cholla project (USA).
}
\seealso{
Other count data: 
\code{\link{birds}},
\code{\link{boves}},
\code{\link{compiegne}},
\code{\link{loire}},
\code{\link{merzbach}},
\code{\link{mississippi}},
\code{\link{zuni}}
}
\concept{count data}
\keyword{datasets}
