% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codex-package.R
\docType{package}
\name{codex-package}
\alias{codex-package}
\alias{_PACKAGE}
\alias{codex}
\title{codex: Datasets for Teaching Archaeology and Paleontology}
\description{
Datasets for teaching quantitative approaches and
    modeling in archaeology and paleontology. This package provides
    several types of data related to broad topics (cultural evolution,
    radiocarbon dating, paleoenvironments, etc.), which can be used to
    illustrate statistical methods in the classroom (multivariate data
    analysis, compositional data analysis, diversity measurement, etc.).
}
\details{
\tabular{ll}{
\strong{Package:} \tab codex \cr
\strong{Type:} \tab Package \cr
\strong{Version:} \tab 1.0.0 \cr
\strong{License:} \tab GPL-3 \cr
\strong{DOI:} \tab \href{https://doi.org/10.5281/zenodo.4476182}{10.5281/zenodo.4476182}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://codex.tesselle.org}
  \item \url{https://github.com/tesselle/codex}
  \item Report bugs at \url{https://github.com/tesselle/codex/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

IRAMAT-CRP2A (UMR 5060)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
