% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\docType{data}
\name{compiegne}
\alias{compiegne}
\title{Compiègne Ceramics}
\format{
A \code{data.frame} with 5 observations (periods) and 16 variables (ceramic types):
\describe{
\item{A}{\code{\link{integer}}: red to white ceramics with fine sized inclusions.}
\item{B}{\code{\link{integer}}: red to white ceramics with medium sized inclusions.}
\item{C}{\code{\link{integer}}: dark ceramics with fine sized inclusions.}
\item{D}{\code{\link{integer}}: dark ceramics with medium sized inclusions.}
\item{E}{\code{\link{integer}}: ceramics close to those of groups B or D, with
similarities to group F.}
\item{F}{\code{\link{integer}}: black, red or beige ceramics with coarse inclusions.}
\item{G}{\code{\link{integer}}: red polished ceramics with fine to medium sized
inclusions.}
\item{H}{\code{\link{integer}}: black polished ceramics with fine sized inclusions.}
\item{I}{\code{\link{integer}}: black polished ceramics with medium sized inclusions.}
\item{J}{\code{\link{integer}}: polished and painted ceramics with fine to medium
sized inclusions.}
\item{K}{\code{\link{integer}}: painted ceramics, similar to those of group A.}
\item{L}{\code{\link{integer}}: painted ceramics, similar to those of group B.}
\item{M}{\code{\link{integer}}: painted ceramics with coarse inclusions.}
\item{N}{\code{\link{integer}}: glazed ceramics.}
\item{O}{\code{\link{integer}}: stamped ceramics.}
\item{P}{\code{\link{integer}}: coated ceramics.}
}
Chronological periods are numbered from the oldest to the most recent
(from 1 to 5).
}
\source{
Lacroix, M. C. (1997). La céramique médiévale du site des Hallettes à
Compiègne (Oise). \emph{Revue archéologique de Picardie}. Numéro spécial,
13(1), 135-168. \doi{10.3406/pica.1997.1945}
}
\usage{
compiegne
}
\description{
A dataset containing the ceramic counts from the Place des Hallettes in
Compiègne (Oise, France). The data are grouped into five periods of about a
century, ranging from the 9th to the 14th century, and sixteen ceramic
types.
}
\seealso{
Other count data: 
\code{\link{birds}},
\code{\link{boves}},
\code{\link{chevelon}},
\code{\link{loire}},
\code{\link{merzbach}},
\code{\link{mississippi}},
\code{\link{zuni}}
}
\concept{count data}
\keyword{datasets}
