% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codex-package.R
\docType{package}
\name{dimensio-package}
\alias{dimensio-package}
\alias{_PACKAGE}
\alias{dimensio}
\title{codex: Datasets for Teaching Archaeology and Palaeontology}
\description{
Datasets for teaching quantitative approaches and
    modeling in archaeology and paleontology. This package provides
    several types of data, relating to typologies, cultural evolution,
    radiocarbon dating and paleoenvironments.
}
\details{
\tabular{ll}{
\strong{Package:} \tab codex \cr
\strong{Type:} \tab Package \cr
\strong{Version:} \tab 1.0.0 \cr
\strong{License:} \tab GPL-3 \cr
\strong{DOI:} \tab \href{https://doi.org/xxx}{xxx}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nfrerebeau/codex}
  \item Report bugs at \url{https://github.com/nfrerebeau/codex/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

IRAMAT-CRP2A (UMR 5060)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
