% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folio-package.R
\docType{package}
\name{folio-package}
\alias{folio-package}
\alias{folio}
\title{folio: Datasets for Teaching Archaeology and Paleontology}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Datasets for teaching quantitative approaches and modeling in archaeology and paleontology. This package provides several types of data related to broad topics (cultural evolution, radiocarbon dating, paleoenvironments, etc.), which can be used to illustrate statistical methods in the classroom (multivariate data analysis, compositional data analysis, diversity measurement, etc.).
}
\details{
\tabular{ll}{
\strong{Version} \tab 1.5.1 \cr
\strong{License} \tab GPL-3 \cr
\strong{CRAN DOI} \tab \doi{10.32614/CRAN.package.folio} \cr
\strong{Zenodo DOI} \tab \doi{10.5281/zenodo.4476182} \cr
}

Archéosciences Bordeaux (UMR 6034)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\seealso{
Useful links:
\itemize{
  \item \url{https://codeberg.org/tesselle/folio}
  \item \url{https://packages.tesselle.org/folio/}
  \item \url{https://tesselle.r-universe.dev/folio}
  \item Report bugs at \url{https://codeberg.org/tesselle/folio/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Frerebeau \email{nicolas.frerebeau@u-bordeaux-montaigne.fr} (\href{https://orcid.org/0000-0001-5759-4944}{ORCID})

Other contributors:
\itemize{
  \item Brice Lebrun (\href{https://orcid.org/0000-0001-7503-8685}{ORCID}) (Logo designer) [artist]
  \item Université Bordeaux Montaigne (\href{https://ror.org/03pbgwk21}{ROR}) [funder]
  \item CNRS (\href{https://ror.org/02feahw73}{ROR}) [funder]
}

}
\keyword{internal}
