% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon.R
\docType{data}
\name{intcal09}
\alias{intcal09}
\title{IntCal09}
\format{
A \code{data.frame} with 3522 observations and 5 variables:
\describe{
\item{calBP}{\code{\link{integer}}: calendar (calibrated) age (year BP).}
\item{age}{\code{\link{integer}}: radiocarbon age (year BP).}
\item{error}{\code{\link{integer}}: radiocarbon error (year BP).}
\item{delta}{\code{\link{numeric}}: isotopic ratio \eqn{\Delta^{14}C}{\Delta14C} (per mil).}
\item{sigma}{\code{\link{numeric}}: error on \code{delta} (per mil).}
}
}
\source{
Reimer, P. J., Baillie, M. G. L., Bard, E., Bayliss, A., Beck, J. W.,
Blackwell, P. G., Bronk Ramsey, C. \emph{et al.} (2009). IntCal09 and Marine09
Radiocarbon age Calibration Curves, 0-50,000 Years Cal BP.
\emph{Radiocarbon}, 51(4): 1111-50. \doi{10.1017/S0033822200034202}.
}
\usage{
intcal09
}
\description{
The IntCal series of radiocarbon calibration curves.
}
\seealso{
Other radiocarbon dating: 
\code{\link{arnold1949}},
\code{\link{intcal13}},
\code{\link{intcal20}}
}
\concept{radiocarbon dating}
\keyword{datasets}
