% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon.R
\docType{data}
\name{intcal20}
\alias{intcal20}
\title{IntCal20}
\format{
A \code{data.frame} with 9501 observations and 5 variables:
\describe{
\item{calBP}{\code{\link{integer}}: calendar (calibrated) age (year BP).}
\item{age}{\code{\link{integer}}: radiocarbon age (year BP).}
\item{error}{\code{\link{integer}}: radiocarbon error (year BP).}
\item{delta}{\code{\link{numeric}}: isotopic ratio \eqn{\Delta^{14}C}{\Delta14C} (per mil).}
\item{sigma}{\code{\link{numeric}}: error on \code{delta} (per mil).}
}
}
\source{
Reimer, P. J., Austin, W. E. N., Bard, E., Bayliss, A., Blackwell, P. G.,
Bronk Ramsey, C., Butzin, M. \emph{et al.} (2020). The IntCal20 Northern
Hemisphere Radiocarbon âge Calibration Curve (0-55 Cal KBP).
\emph{Radiocarbon}, 62(4), 725-757. \doi{10.1017/RDC.2020.41}.
}
\usage{
intcal20
}
\description{
The IntCal series of radiocarbon calibration curves.
}
\seealso{
Other radiocarbon dating: 
\code{\link{arnold1949}},
\code{\link{intcal09}},
\code{\link{intcal13}}
}
\concept{radiocarbon dating}
\keyword{datasets}
