% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-isotopes.R
\docType{data}
\name{lisiecki2005}
\alias{lisiecki2005}
\title{LR04 Benthic Stack}
\format{
A data frame with 3 variables:
\describe{
\item{time}{Calendar (calibrated) age (kilo year).}
\item{delta}{Benthic \eqn{\delta^{18}O}{\delta18O} (per mil).}
\item{error}{Standard error (per mil).}
}
}
\source{
\url{https://www.ncdc.noaa.gov/paleo-search/study/5847}
}
\usage{
lisiecki2005
}
\description{
A global Pliocene-Pleistocene benthic \eqn{\delta^{18}O}{\delta18O} stack.
}
\details{
The LR04 stack spans 5.3 Myr and is an average of 57 globally distributed
benthic \eqn{\delta^{18}O}{\delta18O} records (which measure global ice
volume and deep ocean temperature) collected from the scientific literature.
}
\references{
Lisiecki, L. E. and Raymo, M. E. (2005). A Pliocene-Pleistocene stack of 57
globally distributed benthic d18O records. \emph{Paleoceanography}, 20,
PA1003. \doi{10.1029/2004PA001071}
}
\seealso{
Other isotopic data: 
\code{\link{spratt2016}}
}
\concept{isotopic data}
\keyword{datasets}
