% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\docType{data}
\name{loire}
\alias{loire}
\title{Medieval Ceramics from the Loire Basin}
\format{
A \code{\link{data.frame}} with 226 rows (archaeological sets) and 186
variables. The columns correspond to the counts of 183 ceramic types.
The first three columns provide background information:
\describe{
\item{lower}{Lower bound of the temporal range.}
\item{upper}{Upper bound of the temporal range.}
\item{area}{Geographical area.}
}
}
\source{
\url{https://ceramedvaldeloire.huma-num.fr}
}
\usage{
loire
}
\description{
A dataset containing the ceramic counts from the Loire Basin (France).
}
\references{
Husi, P. (dir.). (2022). \emph{La céramique médiévale et moderne du bassin de la
Loire moyenne, chrono-typologie et transformation des aires culturelles
dans la longue durée (6e—19e s.)}. Suppléments à la revue Archéologique du
Centre de la France, 79.
}
\seealso{
Other count data: 
\code{\link{birds}},
\code{\link{boves}},
\code{\link{chevelon}},
\code{\link{compiegne}},
\code{\link{merzbach}},
\code{\link{mississippi}},
\code{\link{zuni}}
}
\concept{count data}
\keyword{datasets}
