% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\docType{data}
\name{mississippi}
\alias{mississippi}
\title{Mississippi Ceramics}
\format{
A \code{data.frame} with 20 observations and 10 variables (ceramic types):
\describe{
\item{ParkinPunctate}{\code{\link{integer}}.}
\item{BartonKentMPI}{\code{\link{integer}}.}
\item{Painted}{\code{\link{integer}}.}
\item{FortuneNoded}{\code{\link{integer}}.}
\item{RanchIncised}{\code{\link{integer}}.}
\item{WallsEngraved}{\code{\link{integer}}.}
\item{WallaceIncised}{\code{\link{integer}}.}
\item{RhodesIncised}{\code{\link{integer}}.}
\item{VernonPaulApplique}{\code{\link{integer}}.}
\item{HullEngraved}{\code{\link{integer}}.}
}
}
\source{
Lipo, C. P., Madsen, M. E. & Dunnell, R. C. (2015). A
Theoretically-Sufficient and Computationally-Practical Technique for
Deterministic Frequency Seriation. \emph{PLOS ONE}, 10(4), e0124942.
\doi{10.1371/journal.pone.0124942}.
}
\usage{
mississippi
}
\description{
A dataset containing ceramic counts from the Mississippi region.
}
\seealso{
Other count data: 
\code{\link{birds}},
\code{\link{boves}},
\code{\link{chevelon}},
\code{\link{compiegne}},
\code{\link{loire}},
\code{\link{merzbach}},
\code{\link{zuni}}
}
\concept{count data}
\keyword{datasets}
