% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artefacts.R
\docType{data}
\name{munsingen}
\alias{munsingen}
\title{M체nsingen Cemetery}
\format{
A \code{data.frame} with 59 observations (graves) and 70 variables (artefact types).
}
\usage{
munsingen
}
\description{
A dataset of data set of artifact presence/absence for the Celtic
M체nsingen-Rain cemetery (Switzerland).
}
\examples{
heatmap(
  x = as.matrix(munsingen),
  Rowv = NA,
  Colv = NA,
  scale = "none",
  col = c("white", "black")
)
}
\references{
Hodson, F. R. (1968). \emph{The La Tene Cemetery at M체nsingen-Rain}. St채mpfli,
Bern.

Kendall, D. G. (1971). Seriation from abundance matrices. In Hodson,
F. R., Kendall, D. G. and Tautu, P. (eds), \emph{Mathematics in the
Archaeological and Historical Sciences}. Edinburgh University Press,
Edinburgh, 215-232.
}
\seealso{
Other artefact data: 
\code{\link{shipwrecks}}
}
\concept{artefact data}
\keyword{datasets}
