% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-counts.R
\docType{data}
\name{chevelon}
\alias{chevelon}
\title{Chevelon Ground Stone}
\format{
A data frame with 12 rows and 10 variables (ground stone types):
\describe{
\item{BMe}{Basin metate.}
\item{SMe}{Slab metate.}
\item{TMe}{Trough metate.}
\item{IMe}{Indeterminate metate.}
\item{UMa}{Unifacial mano.}
\item{BMa}{Bifacial mano.}
\item{MUHa}{Modified unifacial handstone.}
\item{MBHa}{Modified bifacial handstone.}
\item{UUHa}{Unmodified unifacial handstone.}
\item{UBHa}{Unmodified bifacial handstone.}
}
}
\source{
Reid, J. J. (ed.) (1982). \emph{Cholla Project Archaeology}. Vol. 2.
Archaeological Series 161. Tucson: University of Arizona.
\doi{10.6067/XCV8435710}
}
\usage{
chevelon
}
\description{
A dataset of ground stone artifact counts from the Cholla project (USA).
}
\seealso{
Other count data: 
\code{\link{birds}},
\code{\link{boves}},
\code{\link{compiegne}},
\code{\link{merzbach}},
\code{\link{mississippi}},
\code{\link{zuni}}
}
\concept{count data}
\keyword{datasets}
