% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folio-package.R
\docType{package}
\name{folio-package}
\alias{folio-package}
\alias{folio}
\title{folio: Datasets for Teaching Archaeology and Paleontology}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Datasets for teaching quantitative approaches and modeling in archaeology and paleontology. This package provides several types of data related to broad topics (cultural evolution, radiocarbon dating, paleoenvironments, etc.), which can be used to illustrate statistical methods in the classroom (multivariate data analysis, compositional data analysis, diversity measurement, etc.).
}
\details{
\tabular{ll}{
\strong{Package:} \tab folio \cr
\strong{Type:} \tab Package \cr
\strong{Version:} \tab 1.4.0 \cr
\strong{License:} \tab GPL-3 \cr
\strong{Zenodo:} \tab \doi{10.5281/zenodo.4476182} \cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://packages.tesselle.org/folio/}
  \item \url{https://github.com/tesselle/folio}
  \item Report bugs at \url{https://github.com/tesselle/folio/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
Brice Lebrun \tab \emph{Université Bordeaux Montaigne, France}
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

Archéosciences Bordeaux (UMR 6034)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
F-33607 Pessac cedex\cr
France
}
\keyword{internal}
