% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon.R
\docType{data}
\name{intcal13}
\alias{intcal13}
\title{IntCal13}
\format{
A data frame with 5 variables:
\describe{
\item{calBP}{Calendar (calibrated) age (year BP).}
\item{age}{Radiocarbon age (year BP).}
\item{error}{Radiocarbon error (year BP).}
\item{delta}{Isotopic ratio \eqn{\Delta^{14}C}{\Delta14C} (per mil).}
\item{sigma}{Error on \code{delta} (per mil).}
}
}
\source{
Reimer, P. J., Bard, E. Bayliss, A., Beck, J. W., Blackwell, P. G.,
Bronk Ramsey, C., Buck, C. E. \emph{et al.} (2013). IntCal13 and Marine13
Radiocarbon age Calibration Curves 0-50,000 Years cal BP.
\emph{Radiocarbon}, 55(4): 1869‑87. \doi{10.2458/azu_js_rc.55.16947}.
}
\usage{
intcal13
}
\description{
The IntCal series of radiocarbon calibration curves.
}
\seealso{
Other radiocarbon dating: 
\code{\link{arnold1949}},
\code{\link{intcal09}},
\code{\link{intcal20}}
}
\concept{radiocarbon dating}
\keyword{datasets}
