% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-carbon.R
\docType{data}
\name{intcal20}
\alias{intcal20}
\title{IntCal20}
\format{
A data frame with 5 variables:
\describe{
\item{calBP}{Calendar (calibrated) age (year BP).}
\item{age}{Radiocarbon age (year BP).}
\item{error}{Radiocarbon error (year BP).}
\item{delta}{Isotopic ratio \eqn{\Delta^{14}C}{\Delta14C} (per mil).}
\item{sigma}{Error on \code{delta} (per mil).}
}
}
\source{
Reimer, P. J., Austin, W. E. N., Bard, E., Bayliss, A., Blackwell, P. G.,
Bronk Ramsey, C., Butzin, M. \emph{et al.} (2020). The IntCal20 Northern
Hemisphere Radiocarbon âge Calibration Curve (0-55 Cal KBP).
\emph{Radiocarbon}, 62(4), 725-757. \doi{10.1017/RDC.2020.41}.
}
\usage{
intcal20
}
\description{
The IntCal series of radiocarbon calibration curves.
}
\seealso{
Other calibration curves: 
\code{\link{intcal09}},
\code{\link{intcal13}}
}
\concept{calibration curves}
\keyword{datasets}
