% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemical.R
\docType{data}
\name{kommos}
\alias{kommos}
\title{Transport Jars from Kommos (Crete).}
\format{
A data frame with 22 variables (chemical elements):
\describe{
\item{type}{\code{CJ}: Canaanite jar; \code{EJ}: Egyptian jar;
\code{TSJ}: transport stirrup jar; \code{SNA}: short-necked amphora.}
\item{date}{Chronology (period).}
\item{Sm}{(ppm).}
\item{Lu}{(ppm).}
\item{U}{(ppm).}
\item{Yb}{(ppm).}
\item{As}{(ppm).}
\item{Sb}{(ppm).}
\item{Ca}{(ppm).}
\item{Na}{(ppm).}
\item{La}{(ppm).}
\item{Ce}{(ppm).}
\item{Th}{(ppm).}
\item{Cr}{(ppm).}
\item{Hf}{(ppm).}
\item{Cs}{(ppm).}
\item{Sc}{(ppm).}
\item{Rb}{(ppm).}
\item{Fe}{(ppm).}
\item{Ta}{(ppm).}
\item{Co}{(ppm).}
\item{Eu}{(ppm).}
}
}
\usage{
kommos
}
\description{
Chemical analysis (neutron activation analysis) of 88 Late Bronze Age
transport jars found in excavations at Kommos, Crete.
}
\references{
Day, P. M., Quinn, P. S., Rutter, J. B. & Kilikoglou, V. (2011).
A World of Goods: Transport Jars and Commodity Exchange at the Late Bronze
Age Harbor of Kommos, Crete. \emph{Hesperia}, 80, 511-558.
\doi{10.2972/hesperia.80.4.0511}
}
\concept{chemical data}
\keyword{datasets}
