% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\docType{data}
\name{loire}
\alias{loire}
\title{Medieval Ceramics from the Loire Basin}
\format{
A \code{\link{data.frame}} with 332 rows (archaeological assemblages) and 331
variables. The first five columns provide background information,
the next columns give the MNI of each ceramic types:
\describe{
\item{site}{Name of the archaeological site.}
\item{city}{City.}
\item{area}{Geographical area.}
\item{lower}{Lower bound of the temporal range.}
\item{upper}{Upper bound of the temporal range.}
}
}
\source{
https://ceramedvaldeloire.huma-num.fr/editions/suppl79racf2022/accueil
}
\usage{
loire
}
\description{
A dataset containing the ceramic counts from the Loire Basin (France).
}
\references{
Husi, P. (dir.). (2022). \emph{La céramique médiévale et moderne du bassin de la
Loire moyenne, chrono-typologie et transformation des aires culturelles
dans la longue durée (6e-19e s.)}. Suppléments à la revue Archéologique du
Centre de la France, 79.
}
\seealso{
Other count data: 
\code{\link{birds}},
\code{\link{boves}},
\code{\link{chevelon}},
\code{\link{compiegne}},
\code{\link{merzbach}},
\code{\link{mississippi}},
\code{\link{zuni}}
}
\concept{count data}
\keyword{datasets}
