% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artefacts.R
\docType{data}
\name{shipwrecks}
\alias{shipwrecks}
\title{Mediterranean Shipwrecks}
\format{
A \code{\link{data.frame}} with 1784 rows and 13 variables:
\describe{
\item{name}{Wreck name.}
\item{sea}{Region of the sea where the wreck was discovered.}
\item{country}{Country where the wreck was discovered.}
\item{region}{Region where the wreck was discovered.}
\item{depth_min}{Minimum depth of the wreck (m).}
\item{depth_max}{Maximum depth of the wreck (m).}
\item{depth}{Depth of the wreck (m).}
\item{period}{Period.}
\item{dating}{Dating.}
\item{date_early}{Earliest date.}
\item{date_late}{Latest date.}
\item{origin}{Place of origin.}
\item{destination}{Place of destination.}
}
}
\source{
Strauss, J. (2013). \emph{Shipwrecks Database}. Version 1.0. Accessed 2022-08-13.
URL: \url{http://oxrep.classics.ox.ac.uk/databases/shipwrecks_database/}
}
\usage{
shipwrecks
}
\description{
A dataset of mediterranean shipwrecks.
}
\note{
This dataset contains typos and needs to be normalized.
}
\references{
Parker, A. J. (1992). \emph{Ancient Shipwrecks of the Mediterranean and the Roman
Provinces}. British Archaeological Reports International Series 580. Oxford.

Strauss, E. J. (2007). \emph{Roman Cargoes: Underwater Evidence from the Eastern
Mediterranean}. Doctoral thesis, University College London.
URL: \url{https://discovery.ucl.ac.uk/id/eprint/1349806}.
}
\seealso{
Other artefact data: 
\code{\link{munsingen}}
}
\concept{artefact data}
\keyword{datasets}
