% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chronology.R
\docType{data}
\name{stratigraphy}
\alias{stratigraphy}
\title{Chronostratigraphic Chart}
\format{
A \code{\link{data.frame}} with 5 variables:
\describe{
\item{type}{Unit type ("\code{eon}", "\code{era}", "\code{period}", "\code{series}" or
"\code{stage}"). Precambrian and Hadean are informal units.}
\item{name}{Unit name.}
\item{age}{Numerical age (Ma).}
\item{error}{Error on numerical age (Ma).}
\item{parent}{Parent unit.}
}
}
\source{
\url{https://stratigraphy.org/ICSchart/ChronostratChart2022-02.pdf}
}
\usage{
stratigraphy
}
\description{
The ICS international chronostratigraphic chart (v2022/2).
}
\references{
Cohen, K. M., Finney, S. C., Gibbard, P. L. and Fan, J.-X. (2013).
The ICS International Chronostratigraphic Chart. \emph{Episodes},
36(3): 199-204. \doi{10.18814/epiiugs/2013/v36i3/002}
}
\concept{chronological data}
\keyword{datasets}
